Create hangman.py
If you want to use this you'll have to create a words.py file and populate an list called test with random words.
And then you're free to run, other thing you might want to do is to use an api. I had firtly used an api, but the api wasn't reliable so I gave up and created my own list with words.